Generate Darcula versions in builjar.sh
This is the solution proposed in my comment to #62.
The buildjar.sh script now generates Darcula versions
of both color schemes and includes them in the settings.jar file.
These Darcula versions will be copies of the regular versions
with only the file status colors changed to the colors from
the default Darcula scheme.

I am not knowledgeable enough about Windows scripting
to be able to update the builjar.bat script, which is
therefore removed.

Fixes #41, Fixes #44, Fixes #62, Fixes #66