Define DEFAULT_IDENTIFIER to only have foreground
Since DEFAULT_IDENTIFIER was not defined it inherited its colors
from TEXT, which is where the default background is defined.

This became a problem when identifiers where a part of a template
language snippet, because they would not get the DEFAULT_TEMPLATE_LANGUAGE_COLOR
background but the default background.

We now define DEFAULT_IDENTIFIER with only a foreground color.

Fixes #95 and fixes #101