Change notification background
The notification banner uses a different foreground color for
the default L&F and the Darcula L&F. Therefore we need to use
different notification backgrounds (the foreground doesn't seem
to be customizable).
This is now done when generating the Darcula versions in the
buildjar.sh script.

The default notification background color is also changed to use
base2 (the gutter background) instead of base3 (the editor background).

Fixes #79 and fixes #94